packt-sre-code/packt-sre/py-cars/requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
aws-xray-sdk==2.6.0
boto3==1.13.17
botocore==1.16.17
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dnspython==1.16.0
docutils==0.15.2
eventlet==0.25.2
Flask==1.1.2
future==0.18.2
greenlet==0.4.15
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonpickle==1.4.1
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
pytest-dependency==0.5.1
python-dateutil==2.8.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
